[explorer] refactor: move formatTransferMessage to NEM variant utils
