+++ explorer/frontend/__tests__/utils/common.test.js
@@ -12,7 +12,6 @@ import {
 	formatTransactionCSV,
 	formatTransactionChart,
 	formatTransactionsPerBlock,
-	formatTransferMessage,
 	getRootNamespaceName,
 	nullableValueToText,
 	numberToShortString,
@@ -523,50 +522,6 @@ describe('utils/common', () => {
 
 			// Assert:
 			expect(result).toBe(expectedResult);
-		});
-	});
-
-	describe('formatTransferMessage', () => {
-		it('returns decoded plain text for plain messages', () => {
-			// Arrange:
-			const payload = '48656c6c6f';
-
-			// Act:
-			const result = formatTransferMessage(1, payload);
-
-			// Assert:
-			expect(result).toEqual({
-				type: 'plain',
-				text: 'Hello'
-			});
-		});
-
-		it('returns hex label for plain messages with fe prefix', () => {
-			// Arrange:
-			const payload = 'feABCD1234';
-
-			// Act:
-			const result = formatTransferMessage(1, payload);
-
-			// Assert:
-			expect(result).toEqual({
-				type: 'hex',
-				text: 'HEX: ABCD1234'
-			});
-		});
-
-		it('returns encrypted payload for non-plain messages', () => {
-			// Arrange:
-			const payload = 'A1B2C3D4';
-
-			// Act:
-			const result = formatTransferMessage(2, payload);
-
-			// Assert:
-			expect(result).toEqual({
-				type: 'encrypted',
-				text: 'A1B2C3D4'
-			});
 		});
 	});
